+++ F1BroadcastTools.Client/src/app/features/settings/driver-overrides/driver-overrides.component.html
@@ -1,6 +1,6 @@
-<h1>Player overrides</h1>
+<h1>Driver overrides</h1>
 
-<div class="player-overrides">
+<div class="driver-overrides">
     <table>
         <thead>
             <tr>
