+++ F1BroadcastTools.Client/src/app/features/settings/settings-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from "@angular/core";
 import { Routes, RouterModule } from "@angular/router";
-import { PlayerOverrides } from "./player-overrides/player-overrides.component";
+import { DriverOverrides } from "./driver-overrides/driver-overrides.component";
 
 const routes: Routes = [
     {
         path: '',
-        component: PlayerOverrides
+        component: DriverOverrides
     }
 ];
