+++ F1GameDataParser/State/DriverOverrideState.cs
@@ -1,0 +1,8 @@
+﻿using F1GameDataParser.Models.DriverOverride;
+
+namespace F1GameDataParser.State
+{
+    public class DriverOverrideState : StateBase<DriverOverride>
+    {
+    }
+}
