--- F1GameDataParser/State/PlayerOverrideState.cs
@@ -1,8 +1,0 @@
-﻿using F1GameDataParser.Models.PlayerOverride;
-
-namespace F1GameDataParser.State
-{
-    public class PlayerOverrideState : StateBase<PlayerOverride>
-    {
-    }
-}
